Consolidate vendored dinov2 + delete dead vendored code (2,436 -> 500 lines)
One third_party/dinov2 package now serves moge, depth_anything_v2, and
video_depth_anything. The two source forks differ in live math (moge bakes
SDPA into Attention.forward; pos-embed interpolation differs ~3.5e-6) — the
consolidated ViT keeps both behind explicit constructor knobs, factory-
defaulted per consumer. Deleted as verified-dead: xformers/nested-tensor/
ViT-G/register/chunked/masked/onnx paths, hub weight downloading, dino_head,
sdpa-wrapper + gradient-checkpointing machinery, timeit, unused geometry
helpers, and the whole utils3d shim (its two live functions moved into
moge/utils/geometry_torch, equivalence-checked exactly). Mutable list
defaults -> tuples; predictors pass force_projection=False (nothing reads
points).

Gates: dinov2 forward goldens (both forks, bit-exact, checked in as npz),
moge rrd parity 48 cols atol=0, DAv2 rrd parity 14 cols atol=0, suite 41
passed, lint clean.
